@@ -1,15 +1,15 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, /*OnInit*/ } from '@angular/core';
 
 @Component({
   selector: 'app-mapas',
   templateUrl: './mapas.component.html',
   styleUrls: ['./mapas.component.scss']
 })
-export class MapasComponent implements OnInit {
+export class MapasComponent /*implements OnInit*/ {
 
   constructor() { }
 
-  ngOnInit() {
-  }
+  // ngOnInit() {
+  // }
 
 }
